created the constat.ts file and the hero section ts file
